IAM for authenticating to Sysdig Service
The REST APIs / Python libraries are updated for team support when using the IBM IAM authentication.

Earlier when running APIs using IAM authentication default to the Monitoring team since the same API token is used regardless of the team to which you have permissions.  The API has been expanded to allow the caller to indicate which team ( optionally ) they are attempting to access.